upgrade boot and move to H2

+++ GetItFreeService/src/main/java/com/getItFree/exception/ResourceConflictException.java
@@ -5,18 +5,18 @@ import java.math.BigInteger;
 public class ResourceConflictException extends RuntimeException {
 
     private static final long serialVersionUID = 1791564636123821405L;
-    private BigInteger resourceId;
+    private Long resourceId;
 
-    public ResourceConflictException(BigInteger resourceId, String message) {
+    public ResourceConflictException(Long resourceId, String message) {
         super(message);
         this.setResourceId(resourceId);
     }
 
-    public BigInteger getResourceId() {
+    public Long getResourceId() {
         return resourceId;
     }
 
-    public void setResourceId(BigInteger resourceId) {
+    public void setResourceId(Long resourceId) {
         this.resourceId = resourceId;
     }
 }
